NeuroMonika/requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
blis==0.7.4
bs4==0.0.1
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
cmudict==1.0.2
command-not-found==0.3
cymem==2.0.5
dbus-python==1.2.16
distro-info==1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
fake-headers==1.0.2
googletrans==4.0.0rc1
gyp==0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.0.1
language-selector==0.1
libretranslatepy==2.1.1
logging==0.4.9.6
lxml==4.6.3
lytspel==2.0.4
MarkupSafe==2.0.1
multidict==5.1.0
murmurhash==1.0.5
netifaces==0.10.9
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pathy==0.6.0
preshed==3.0.5
pronouncing==0.2.0
pydantic==1.8.2
PyGObject==3.38.0
pyparsing==2.4.7
pyTelegramBotAPI==4.0.0
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
PyYAML==5.3.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.9
tornado==6.1
tqdm==4.62.2
translate==3.6.1
typer==0.3.2
typing-extensions==3.10.0.2
tzlocal==3.0
ubuntu-advantage-tools==27.2
ufw==0.36
urllib3==1.26.6
wasabi==0.8.2
webencodings==0.5.1
yarl==1.6.3
